[PYTHON] Check for HINT in map, fix circular import
