+++ tools/benchmark_tool/openvino/tools/benchmark/parameters.py
@@ -154,4 +154,5 @@ def parse_args():
                            "Example: -imean data[255,255,255],info[255,255,255]\n")
     parsed_args = parser.parse_args()
 
+
     return parsed_args
